+++ scripts/installDependencies.js
@@ -1,9 +1,12 @@
 import fs from "fs-extra";
 import path from "path";
+import sass from "sass";
 
 import {cp} from "./common/file-operations.js";
 import {info, error} from "./common/custom-console.js";
 import {projectRootDir} from "./projectRootDir.js";
+import {execSync} from "./common/custom-child-process.js";
+import * as child_process from "child_process";
 
 const fontPath = path.join(projectRootDir, './webextension/assets/fonts/'),
 	jsLib = path.join(projectRootDir, './webextension/lib/')
@@ -25,6 +28,29 @@ if (!exist_jsLib) {
 	error("JS lib folder not found!");
 	process.exit(1);
 } else {
+	/*info("Generating `manifest.json`...");
+	const manifestJsonPath = path.join(projectRootDir, "/webextension/manifest.json");
+	if (!fs.existsSync(manifestJsonPath)) {
+		const privateKeyPath = path.join(projectRootDir, "/key.pem");
+		if (!fs.existsSync(privateKeyPath)) {
+			child_process.execSync('openssl genrsa 2048 | openssl pkcs8 -topk8 -nocrypt -out key.pem', {
+				encoding: 'utf8',
+				cwd: projectRootDir,
+			});
+		}
+
+		const _manifestJson = fs.readJsonSync(`${manifestJsonPath}.dist`);
+		_manifestJson.key = child_process.execSync('openssl rsa -in key.pem -pubout -outform DER | openssl base64 -A', {
+			encoding: 'utf8',
+			cwd: projectRootDir,
+		});
+		fs.writeJsonSync(manifestJsonPath, _manifestJson, {
+			encoding: 'utf-8',
+			spaces: "\t",
+			EOL: "\n",
+		});
+	}*/
+
 	info("Copying twing...");
 	_cp("./node_modules/twig/twig.min.js", jsLib);
 	_cp("./node_modules/twig/twig.min.js.map", jsLib);
@@ -35,8 +61,16 @@ if (!exist_jsLib) {
 	info("Copying i18next-http-backend...");
 	_cp("./node_modules/i18next-http-backend/i18nextHttpBackend.js", jsLib);
 
-	info("Copying MaterialIcons (marella/material-icons)...");
-	_cp("./node_modules/material-icons/iconfont/material-icons.woff2", path.normalize(`${fontPath}/MaterialIcons-Regular.woff2`));
+	info("Copying MaterialIcons (material-symbols)...");
+	_cp("./node_modules/material-symbols/material-symbols-outlined.woff2", path.normalize(`${fontPath}/material-symbols-outlined.woff2`));
+	_cp("./node_modules/material-symbols/material-symbols-rounded.woff2", path.normalize(`${fontPath}/material-symbols-rounded.woff2`));
+	_cp("./node_modules/material-symbols/material-symbols-sharp.woff2", path.normalize(`${fontPath}/material-symbols-sharp.woff2`));
+	fs.writeFileSync(
+		path.join(projectRootDir, './webextension/assets/fonts/material-symbols.css'), sass.compile("./node_modules/material-symbols/index.scss").css
+			.replace(/ {2}/g, '\t')
+			.replace(/(font-family: "Material Symbols \w+?";)/g, '/*noinspection CssNoGenericFontName*/\n\t$1 /* stylelint-disable-line font-family-no-missing-generic-family-keyword */'),
+		{ encoding: 'utf-8' }
+	);
 
 	info("Copying ip-regex...");
 	_cp("./node_modules/ip-regex/index.js", path.normalize(`${jsLib}/ip-regex.js`));