+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.1.0
+* i: Move from `Material Icons` to `Material Symbols`
+* Fix: Better Vivaldi detection for WebSocket (using speed dial and bookmark bar properties)
+
 # 2.0.2
 * Fix: Syntax error in Firefox